requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==35.0.0
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.7
django-allauth==0.44.0
django-appconf==1.0.5
django-debug-toolbar==3.2.2
django-filter==21.1
django-imagekit==4.0.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.2
Markdown==3.3.4
oauthlib==3.1.1
pilkit==2.0
Pillow==8.3.2
pycparser==2.20
PyJWT>=2.1.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
